Draft - correctly do not change labVBox ID with path lock.

+++ ubiquitous_bash.sh
@@ -8124,6 +8124,12 @@ _vboxlabSSH() {
 	ssh -q -F "$scriptLocal"/vblssh -i "$scriptLocal"/id_rsa "$1"
 }
 
+_labVBox_migrate() {
+	false
+}
+
+
+
 _prepare_instance_vbox() {
 	_prepare_vbox "$instancedVirtDir"
 }
@@ -13148,6 +13154,8 @@ _unset_vbox() {
 _reset_vboxLabID() {
 	[[ "$VBOX_ID_FILE" == "" ]] && _messagePlain_bad 'blank: VBOX_ID_FILE' && return 1
 	
+	[[ "$ub_VBoxLab_prepare" == "true" ]] && return 0
+	
 	rm -f "$VBOX_ID_FILE" > /dev/null 2>&1
 	
 	[[ -e "$VBOX_ID_FILE" ]] && _messagePlain_bad 'fail: VBOX_ID_FILE exists' && return 1
@@ -13212,7 +13220,9 @@ _prepare_vbox() {
 }
 
 _prepare_lab_vbox() {
+	export ub_VBoxLab_prepare='true'
 	_prepare_vbox "$scriptLocal"
+	export ub_VBoxLab_prepare='false'
 }
 #_prepare_lab_vbox
